Tooling: Add auto-generation of marching squares
Make it so that every "ms-*" src tile is interpreted to be a marching
squares background.

We now deal with with 19 out of 21 possible cases.

First we have the 15 original marching square patterns plus 2 alternate
patterns for 5 and 10. The second variant however complicates teh
multi-color marching square tiling, so we use 1 for now.

Using multi-color marching squares further complicates things. Imagine
the following neighorhood of cells

    1122
    1122
    3344
    3344

for the square in the center containing all tiles 1-4 the normally
generated pattern is wrong. Each cell finds itself being the only right
tile color and thus assumes in a binary marching squares setting, that
the other color will will the rest of the tile but a tiny triangle in
the corner. But since all four only draw that circle we get a diamond
shape in the center that is not initialized with tiles.

To solve that we recognize these cases by the fact that they happen when
only one bit is set (cases 1,2, 4 and 8) where the diagonals across are
not the same tile color.

For these cases we introduce 4 new marching squares tiles, which is a
tiny square filling a whole 1/4 corner of a tile, rotated four times.