Implement macro path planning
First experiment with just a grid-based a-* path planning proved to be
prohibitively slow. Even with "just" 2048x2048 the complexity is just
too much.

So we create a navigation mesh and run a macro a-* on it that finds
low-detail paths for world distances. Then we use the grid based a-*
to solve the paths an the much shorter thus much less complex segments
of the macro path. We also can do that incrementally.

Navigation Mesh
===============

To construct the navigation mesh, we first turn our grids into n-sided
polygons with the marching square algorithm using an algorithm varianted
that directly creates the correctly winded n-sided polygons.

Then we use Douglas-Peucker to simplify the the n-sided polygons and
remove very small ones.

The vertices of the simplified n-sided polygons are then fed into a
Delaunay triangulization. This produces as nice triangles as possible.

Then we need fix the contours of the original n-sided polygons by
"flipping" all triangles that share an edge that crosses one of the
original n-sided polygon edges.

For every two wrong triangles [A,B,C] [C,B,D] with a shared edge BC
we replace those two triangles with [B,A,D] [C,D,A].

To remove all obstacle triangles from the mesh we use a mask which we
draw from the simplified polygons using an even-odd xor filling.

The mask contains a perfectly continuous of the simplified polygons so
we can then just take the centroid of our navigation mesh triangles to
determine whether the triangle is obstacle or not.

Workers
=======

We use the webpack "worker-loader" to create WebWorkers for the
long-running tasks like map generation and path planning.

This keeps the render and animation loops going while the worker thread
can do a round-robin of path segments in the background.

We centralize all current tasks into a Services module wrapping the raw
web worker and providing a promise-based query-reply mechanism.