Implement Local navigation
Implement local a-* based on the actual grid that solves the segments of
the macro path solver.

This version animates the local path segments on local.html.

White tiles are walkable with different walkabiltiy.

Black tiles are obstacles.

The crossed out tiles are the closedList, the openList is rendered
greenish-blue.